Shop-Assistant using voice command
This project works as a assistant for a shop where user gives item name as command, and the code search for the specified item from database. If the item is present it will request for the quantity, If the specified quantity meets the presence of stock, the total item price will be calculated and bill will be generated.